Fixed int/long comparison problem in method "private boolean setProperty(String propertyName, long propertyValueNew)"

+++ src/rtf/com/lowagie/text/rtf/parser/properties/RtfProperty.java
@@ -368,7 +368,7 @@ public class RtfProperty {
 		if(propertyName == null) return false;
 		Object propertyValueOld = getProperty(propertyName);
 		if(propertyValueOld instanceof Long) {
-			int valueOld = ((Integer)propertyValueOld).intValue();
+			long valueOld = ((Long)propertyValueOld).longValue();
 			if (valueOld==propertyValueNew) return true;
 		} 
 		beforeChange(propertyName);
